Convert reST files into markdown properly
Obs.: Unfortunately it is not possible to convert docs/license.rst into
markdown because the toctree directive needs a title and the original
file is a TXT without one. So both steps are needed: (a) define a title
and (b) use a special sphinx/reST directive (not available in
recommonmark) to directly include the contents of another file.